requirements.txt
Python 3.10.12
dj-database-url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
psycopg2==2.9.9
python-decouple==3.8
